Add files via upload
This is not just a single page, the navigation buttons will work and linked to one another as shown in the video posted in my LinkedIn profile. All the menu buttons will work. From home page you can go through the portfolio shoot that opens the login page and then submit button extends to another page. You can check the main contents of this page. And Order of the above files :
1. Landing page 
2. Portfolio 
3. About us
4. Services
5. Contact us
6. Registration form
7. Submission confirmation page